backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Google/YouTube
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0

# OpenAI
openai==1.10.0

# Audio processing
yt-dlp>=2024.3.10

# Transcripts
youtube-transcript-api==0.6.2

# PDF generation
fpdf2==2.7.8
Pillow==10.2.0

# Google Cloud Storage
google-cloud-storage==2.14.0

# Background jobs
redis==5.0.1
rq==1.16.0

# Utilities
python-dotenv==1.0.0
email-validator>=2.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
requests==2.31.0
